libs/AI Agent 与大模型应用开发实战手册/第七章 垂直场景实战项目/7.3 项目三：数据分析 Agent（Text-to-SQL）/requirement.txt
openai==1.51.0
anthropic==0.34.0
litellm==1.49.0
pandas==2.2.3
plotly==5.24.0
pydantic==2.9.2
tiktoken==0.7.0
numpy==2.1.1
sqlparse==0.5.1
kaleido==0.2.1
tenacity==9.0.0
python-dotenv==1.0.1